Remove linux, macos, windows ARM
I cannot figure out how to get linux with unity working. The Unity OpenXR plugin only supports Windows+MacOS and gives a explicit error when building. Valve's OpenVR plugin can build and run after some hacks but fails to get past the screen where SteamVR recognises a game starting.

SteamVR support for MacOS was dropped. There are legacy builds so it may still be possible to build for MacOS. However I cannot test this because I have a quest. It is likely this will work for a valve headset.